expanded system module to subpackage structure
